+++ src/MillionaireGame/Forms/Options/DatabaseSettingsDialog.cs
@@ -1,5 +1,6 @@
 using MillionaireGame.Core.Settings;
 using MillionaireGame.Core.Helpers;
+using Microsoft.Data.SqlClient;
 
 namespace MillionaireGame.Forms.Options;
 
@@ -97,9 +98,7 @@ public partial class DatabaseSettingsDialog : Form
             // Try to open connection
             await Task.Run(() =>
             {
-#pragma warning disable CS0618 // Type or member is obsolete
-                using var connection = new System.Data.SqlClient.SqlConnection(connectionString);
-#pragma warning restore CS0618
+                using var connection = new SqlConnection(connectionString);
                 connection.Open();
             });